wakee_lead/requirements.txt
torch
torchvision
numpy==1.26.4
pandas
scikit-learn
matplotlib
tqdm
opencv-python==4.7.0.72
dotenv
pillow
onnxruntime==1.20.1
langchain_core
langchain_mistralai